Serving/requirements.txt
fastapi
uvicorn
Pillow
numpy
scikit-learn==1.2.2
python-multipart
pandas